backend/requirements.txt
Django
django-ninja
djangorestframework
psycopg2-binary  
python-decouple
django-cors-headers
django-filter